+++ src/NetCoreForce.Client/JsonSerializer.cs
@@ -1,6 +1,7 @@
 using System;
 using Newtonsoft.Json;
 using NetCoreForce.Client.Serializer;
+using System.Collections.Generic;
 
 namespace NetCoreForce.Client
 {
@@ -12,8 +13,10 @@ namespace NetCoreForce.Client
         /// </summary>
         /// <param name="inputObject">Object to serialize</param>
         /// <param name="indented">use indented formatting, usually for readability</param>
+        /// <param name="fieldsToNull">A list of properties that should be set to null, but inclusing the null values in the serialized output</param>
+        /// <param name="ignoreNulls">Use with caution. By default null values are not serialized, this will serialize all explicitly nulled or missing properties as null</param>  
         /// <returns>JSON string</returns>
-        public static string SerializeComplete(object inputObject, bool indented)
+        public static string SerializeComplete(object inputObject, bool indented, List<string> fieldsToNull = null, bool ignoreNulls = true)
         {
             Formatting formatting = Formatting.None;
             if(indented)
@@ -25,7 +28,8 @@ namespace NetCoreForce.Client
                    formatting,
                    new JsonSerializerSettings
                    {
-                       NullValueHandling = NullValueHandling.Ignore,
+                       NullValueHandling = ignoreNulls ? NullValueHandling.Ignore : NullValueHandling.Include,
+                       ContractResolver = new NullableContractResolver(fieldsToNull),
                        DateFormatString = DateFormats.FullDateFormatString
                    });
 
@@ -36,15 +40,17 @@ namespace NetCoreForce.Client
         /// Serializes an object into JSON for SObject updates, using the UpdateableContractResolver
         /// </summary>
         /// <param name="inputObject">Object to serialize</param>
-        /// <returns>JSON string, unformatted</returns>
-        public static string SerializeForUpdate(object inputObject)
+        /// <param name="fieldsToNull">A list of properties that should be set to null, but inclusing the null values in the serialized output</param>
+        /// <param name="ignoreNulls">Use with caution. By default null values are not serialized, this will serialize all explicitly nulled or missing properties as null</param>        
+        /// <returns></returns>
+        public static string SerializeForUpdate(object inputObject, List<string> fieldsToNull = null, bool ignoreNulls = true)
         {
             string serializedJson = JsonConvert.SerializeObject(inputObject,
                    Formatting.None,
                    new JsonSerializerSettings
                    {
-                       NullValueHandling = NullValueHandling.Ignore,
-                       ContractResolver = new UpdateableContractResolver(),
+                       NullValueHandling = ignoreNulls ? NullValueHandling.Ignore : NullValueHandling.Include,
+                       ContractResolver = new UpdateableContractResolver(fieldsToNull),
                        DateFormatString = DateFormats.FullDateFormatString
                    });
 
@@ -56,15 +62,17 @@ namespace NetCoreForce.Client
         /// Includes the SObject ID for calls that require it
         /// </summary>
         /// <param name="inputObject">Object to serialize</param>
+        /// <param name="fieldsToNull">A list of properties that should be set to null, but inclusing the null values in the serialized output</param>
+        /// <param name="ignoreNulls">Use with caution. By default null values are not serialized, this will serialize all explicitly nulled or missing properties as null</param>  
         /// <returns>JSON string, unformatted</returns>
-        public static string SerializeForUpdateWithObjectId(object inputObject)
+        public static string SerializeForUpdateWithObjectId(object inputObject, List<string> fieldsToNull = null, bool ignoreNulls = true)
         {
             string serializedJson = JsonConvert.SerializeObject(inputObject,
                    Formatting.None,
                    new JsonSerializerSettings
                    {
-                       NullValueHandling = NullValueHandling.Ignore,
-                       ContractResolver = new UpdateableWithIdContractResolver(),
+                       NullValueHandling = ignoreNulls ? NullValueHandling.Ignore : NullValueHandling.Include,
+                       ContractResolver = new UpdateableWithIdContractResolver(fieldsToNull),
                        DateFormatString = DateFormats.FullDateFormatString
                    });
 
@@ -75,15 +83,17 @@ namespace NetCoreForce.Client
         /// Serializes an object into JSON for SObject creation, using the CreateableContractResolver
         /// </summary>
         /// <param name="inputObject">Object to serialize</param>
+        /// <param name="fieldsToNull">A list of properties that should be set to null, but inclusing the null values in the serialized output</param>
+        /// <param name="ignoreNulls">Use with caution. By default null values are not serialized, this will serialize all explicitly nulled or missing properties as null</param>  
         /// <returns>JSON string, unformatted</returns>
-        public static string SerializeForCreate(object inputObject)
+        public static string SerializeForCreate(object inputObject, List<string> fieldsToNull = null, bool ignoreNulls = true)
         {
             string serializedJson = JsonConvert.SerializeObject(inputObject,
                    Formatting.None,
                    new JsonSerializerSettings
                    {
-                       NullValueHandling = NullValueHandling.Ignore,
-                       ContractResolver = new CreateableContractResolver(),
+                       NullValueHandling = ignoreNulls ? NullValueHandling.Ignore : NullValueHandling.Include,
+                       ContractResolver = new CreateableContractResolver(fieldsToNull),
                        DateFormatString = DateFormats.FullDateFormatString
                    });