+++ src/NetCoreForce.Client/Serializer/NullableContractResolver.cs
@@ -1,4 +1,4 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using System.Reflection;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
@@ -6,23 +6,17 @@ using NetCoreForce.Client.Attributes;
 
 namespace NetCoreForce.Client.Serializer
 {
-    public class CreateableContractResolver : DefaultContractResolver
+    public class NullableContractResolver : DefaultContractResolver
     {
         List<string> _fieldsToNull;
-        public CreateableContractResolver(List<string> fieldsToNull = null)
+        public NullableContractResolver(List<string> fieldsToNull = null)
         {
             _fieldsToNull = fieldsToNull;
         }
-
+        
         protected override JsonProperty CreateProperty(MemberInfo member, MemberSerialization serialization)
         {
             var property = base.CreateProperty(member, serialization);
-
-            var createableAttribute = member.GetCustomAttribute<CreateableAttribute>();
-            if (createableAttribute != null)
-            {
-                property.ShouldSerialize = x => createableAttribute.Createable;
-            }
 
             if (_fieldsToNull != null && _fieldsToNull.FindIndex(x => x.Equals(property.PropertyName, System.StringComparison.OrdinalIgnoreCase)) != -1)
             {