+++ src/NetCoreForce.Client/Serializer/UpdateableWithIdContractResolver.cs
@@ -1,3 +1,5 @@
+using System;
+using System.Collections.Generic;
 using System.Reflection;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
@@ -7,6 +9,12 @@ namespace NetCoreForce.Client.Serializer
 {
     public class UpdateableWithIdContractResolver : DefaultContractResolver
     {
+        List<string> _fieldsToNull;
+        public UpdateableWithIdContractResolver(List<string> fieldsToNull = null)
+        {
+            _fieldsToNull = fieldsToNull;
+        }
+
         protected override JsonProperty CreateProperty(MemberInfo member, MemberSerialization serialization)
         {
             var property = base.CreateProperty(member, serialization);
@@ -17,9 +25,14 @@ namespace NetCoreForce.Client.Serializer
                 property.ShouldSerialize = x => updateableAttribute.Updateable;
             }
 
-            if (property.PropertyName == "id")
+            if (string.Equals(property.PropertyName, "id", StringComparison.OrdinalIgnoreCase))
             {
                 property.ShouldSerialize = x => true;
+            }
+
+            if (_fieldsToNull != null && _fieldsToNull.FindIndex(x => x.Equals(property.PropertyName, System.StringComparison.OrdinalIgnoreCase)) != -1)
+            {
+                property.NullValueHandling = NullValueHandling.Include;
             }
 
             return property;